% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-cell-check.R
\name{check_for_small_size_cells}
\alias{check_for_small_size_cells}
\title{Check for Small Cells}
\usage{
check_for_small_size_cells(passed_table, small_size = 6, print = TRUE,
  table_type = "TableOne")
}
\arguments{
\item{passed_table}{The object outputted by the CreateTableOne function of the tableone package. The documentation is available here
https://cran.r-project.org/web/packages/tableone/index.html.}

\item{small_size}{What value constitutes a small size cell. Default value is 6.}

\item{print}{If TRUE prints the small_size metadata in a human readable format}

\item{table_type}{Specifies the type of the table that is passed to the function}
}
\value{
The passed_table object with a new object in the Metadata object called small_cells. small_cells is a dataframe with 2 columns
variable_name and factors and the rows are all the categorical variables whose one or more factors have small cells.
}
\description{
Small Cells Check checks a given table for small sells then adds a small_cells table to the MetaData of the table object
}
\details{
Checks the categorical table within the TableOne param (CatTable field) for small cells. A small cell is a category where
the number of people in the category (n) is less than the value specified by the small_size param. The freq field within
each variable has the n values.
}
