% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-one.R
\name{table_one_small_cells}
\alias{table_one_small_cells}
\title{Small Cells Check for TableOne}
\usage{
table_one_small_cells(table_one, small_size = 6)
}
\arguments{
\item{table_one}{The object outputted by the CreateTableOne function of the tableone package. The documentation is available here
https://cran.r-project.org/web/packages/tableone/index.html.}

\item{small_size}{What value constitutes a small size cell. Default value is 6.}
}
\value{
data frame with 2 columns variable_name and factors. This only adds the variables that contain small cells for easy identification. It
returns an empty table when no small cells are present
}
\description{
Check for Small Cells inside a TableOne format Table
}
\details{
Check The CatTable list for all possible small cells this also checks for all levels similar to showAllLevels in the CatTable from TableOne
documentation available here:
}
